Refactor; begin moving writer code to standalone
This is to create better abstraction between the choice of key value
store and the index generation code.
